feedhandler/requirements.txt
pandas>=1.5.0
pydantic==1.10.13
APScheduler>=3.8.0
kafka-python>=2.0.2
futu-api
fastapi